+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 
 
+## [0.4.2](https://github.com/Blobfolio/utc2k/releases/tag/v0.4.2) - TBD
+
+### Added
+
+* `utc2k::year` (fetch current year)
+
+
+
+### Changed
+
+* `Utc2k::parse_time_str` now accepts any `AsRef<[u8]>`
+* `Utc2k::from_datetime_str` now accepts any `AsRef<[u8]>`
+* `Utc2k::from_date_str` now accepts any `AsRef<[u8]>`
+* impl `TryFrom<&[u8]>` for `Utc2k` and `FmtUtc2k`
+
+
+
 ## [0.4.1](https://github.com/Blobfolio/utc2k/releases/tag/v0.4.1) - 2022-03-20
 
 ### Added